Restructure project files, add namespaces to composer.json, rename utils to http, removed require_once instructions
