+++ src/Core/Router.php
@@ -4,8 +4,6 @@ namespace Elephantino\Core;
 
 use Elephantino\Traits\RoutesTrait;
 
-require_once "{$rootPath}/src/traits/RoutesTrait.php";
-
 /**
  * Router class is supposed to be used to split endpoints
  * in the app to improve code readability
